Improve pipeline functionality & Windows PowerShell
Previously, a bunch of functions were simply adding the results from the
process block to a collection, which got returned in the end block. Of
course, this was a bad idea, as it defeats the purpose of using the
pipeline in the first place. Additionally, changing the way the few
generic lists are constructed should move the module much closer to
being Windows PowerShell friendly.